// src/api/sourceFotovoltaici/sourceFotovoltaici.entity.ts
export type SourceFotovoltaici = {
    DeviceType: string;
    Name: string;
    Description: string;
    ServerName: string;
    TopicName: string;
    AddressModBus: string;
    AddressDeviceId?: string | null;
    AddressIp?: string | null;
    Coef: string;
    Offset: string;
    LogEnabled: string;
    AlEnabled: string;
    AlBool: string;
    MemTag: string;
    MbsTcpEnabled: string;
    MbsTcpFloat: string;
    SnmpEnabled: string;
    RTLogEnabled: string;
    AlAutoAck: string;
    ForceRO: string;
    SnmpOID: string;
    AutoType: string;
    AlHint?: string | null;
    AlHigh: string;
    AlLow: string;
    AlTimeDB: string;
    AlLevelDB: string;
    IVGroupA: string;
    IVGroupB: string;
    IVGroupC: string;
    IVGroupD: string;
    PageId: string;
    RTLogWindow: string;
    RTLogTimer: string;
    LogDB: string;
    LogTimer: string;
    AlLoLo?: string | null;
    AlHiHi?: string | null;
    MbsTcpRegister: string;
    MbsTcpCoef: string;
    MbsTcpOffset: string;
    EEN?: string | null;
    ETO?: string | null;
    ECC?: string | null;
    ESU?: string | null;
    EAT?: string | null;
    ESH?: string | null;
    SEN?: string | null;
    STO?: string | null;
    SSU?: string | null;
    TEN?: string | null;
    TSU?: string | null;
    FEN?: string | null;
    FFN?: string | null;
    FCO?: string | null;
    KPI?: string | null;
    UseCustomUnit: string;
    Type: string;
    Unit?: string | null;
    AlStat: string;
    ChangeTime: string;
    TagValue: string;
    TagQuality: string;
    AlType: string;
}